Move docopy out of the installer
It doesnt neccesary has to be attached to it and can be reused by other
components
