+++ pkg/utils/copy_test.go
@@ -1,26 +1,9 @@
-/*
-Copyright © 2021 SUSE LLC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-
-package action
+package utils
 
 import (
 	"fmt"
 	. "github.com/onsi/gomega"
-	"github.com/rancher-sandbox/elemental-cli/pkg/types/v1"
+	v1 "github.com/rancher-sandbox/elemental-cli/pkg/types/v1"
 	"io/ioutil"
 	"os"
 	"testing"
@@ -42,9 +25,7 @@ func TestDoCopyEmpty(t *testing.T) {
 		CloudInit: "",
 	}
 
-	install := NewInstallAction(cfg)
-
-	err = install.doCopy()
+	err = DoCopy(cfg)
 	Expect(err).To(BeNil())
 }
 
@@ -68,9 +49,8 @@ func TestDoCopy(t *testing.T) {
 		CloudInit: "",
 	}
 
-	install := NewInstallAction(cfg)
 
-	err = install.doCopy()
+	err = DoCopy(cfg)
 	Expect(err).To(BeNil())
 
 	filesDest, err := ioutil.ReadDir(d)
@@ -108,9 +88,7 @@ func TestDoCopyEmptyWithCloudInit(t *testing.T) {
 		CloudInit: cloudInit.Name(),
 	}
 
-	install := NewInstallAction(cfg)
-
-	err = install.doCopy()
+	err = DoCopy(cfg)
 	Expect(err).To(BeNil())
 	filesDest, err := ioutil.ReadDir(fmt.Sprintf("%s/oem", d))
 	destNames := getNamesFromListFiles(filesDest)
